refactor: organize middleware
