+++ server/routes/closet/closet-routes.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import controller from './closet-controller.js';
-import { validateRequestBody as validate, withTypes } from '../../utils/validation.js';
+import { validateRequestBody as validate } from '../../middleware/validation-middleware.js';
+import { withTypes } from '../../utils/validation.js';
 import {
 	GearClosetItemUpdateSchema,
 	GearClosetItemMoveSchema,
