+++ server/routes/pack/pack-scraper.ts
