+++ server/routes/profile-settings/profile-settings-routes.ts
@@ -2,7 +2,8 @@ import express from 'express';
 import controller from './profile-settings-controller.js';
 import { s3UploadPhoto } from '../../utils/s3.js';
 import { uploadRateLimit } from '../../config/rate-limiting.js';
-import { validateRequestBody as validate, withTypes } from '../../utils/validation.js';
+import { validateRequestBody as validate } from '../../middleware/validation-middleware.js';
+import { withTypes } from '../../utils/validation.js';
 import { 
 	ProfileSettingsUpdateSchema, 
 	UsernameUpdateSchema,
