+++ server/routes/user-settings/user-settings-routes.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import controller from './user-settings-controller.js';
-import { validateRequestBody as validate, withTypes } from '../../utils/validation.js';
+import { validateRequestBody as validate } from '../../middleware/validation-middleware.js';
+import { withTypes } from '../../utils/validation.js';
 import { UserSettingsUpdateSchema, UserSettingsUpdate } from './user-settings-schemas.js';
 
 const router = express.Router();
